Create a server document from what is happening on port 5006 for internal oceans and pass arguments to panel webapp that can be accessed via pn.state.curdoc.session_context.request.arguments
