+++ netcdf_editor_app/db.py
@@ -4,6 +4,8 @@ import click
 from flask import current_app, g
 from flask.cli import with_appcontext
 
+import xarray as xr
+import os
 
 def get_db():
     if 'db' not in g:
@@ -29,6 +31,30 @@ def init_db():
     with current_app.open_resource('db_schema.sql') as f:
         db.executescript(f.read().decode('utf8'))
 
+def load_file(_id):
+    # Get filename
+    print(_id)
+    filepath = get_file_path(_id)
+    # Load file
+    return xr.open_dataset(filepath)
+
+
+def get_file_path(_id, full=True):
+    db = get_db()
+    filepath = db.execute(
+        'SELECT filepath FROM data_files WHERE id = ?', (str(_id), )
+    ).fetchone()['filepath']
+    if not full:
+        return filepath
+    return os.path.join(current_app.instance_path, filepath)
+
+
+def get_lon_lat_names(_id):
+    db = get_db()
+    return db.execute(
+        'SELECT longitude, latitude FROM data_files WHERE id = ?', (str(_id), )
+    ).fetchone()
+
 
 @click.command('init-db')
 @with_appcontext
@@ -41,3 +67,4 @@ def init_db_command():
 def init_app(app):
     app.teardown_appcontext(close_db)
     app.cli.add_command(init_db_command)
+
